move to using gRPC directly
Moves otel-cli to use gRPC directly instead of using the wrapper code
from the OTel collector.

Along the way, it made sense to change span-generating commands to start
the connection independent of sending the span, so connection can happen
in the background while other work is being done.

Also in testing things out manually, I noticed that --timeout on a
process that exceeds the timeout is broken, and things can hang. Put
some hax in to make it work for now, and will come back later to fix it
right, probably by adding --kill-timeout or something like that to make
it explicit the process otel-cli runs may be killed.