custom OTLP client (#205)
This PR moves otel-cli from using the opentelemetry-collector backend for
sending spans to using its own clients directly on top of gRPC and net/http.
Doing so removes some side-effects that the otel collector has, that are
undesirable in otel-cli, such as direct envvar reading. otel-cli's needs are also
simpler than the collector so it doesn't need the layers of abstraction. As a
result, otel-cli's errors are now directly handled and clearer. The binary is
also about 200kb smaller.

Squashed commitlog follows:

* stub out files, no real code yet

* move to using gRPC directly

Moves otel-cli to use gRPC directly instead of using the wrapper code
from the OTel collector.

Along the way, it made sense to change span-generating commands to start
the connection independent of sending the span, so connection can happen
in the background while other work is being done.

Also in testing things out manually, I noticed that --timeout on a
process that exceeds the timeout is broken, and things can hang. Put
some hax in to make it work for now, and will come back later to fix it
right, probably by adding --kill-timeout or something like that to make
it explicit the process otel-cli runs may be killed.

* move grpc client code to its own file

* move http client to its own file

* implement HTTP client

Looks like first pass works, but needs some work on error handling.

Also add the new grpc file because I forgot to add on last commit. This
is all getting squashed at the end anyways.

* remove some carried over grpc bits from http client

* all functional tests pass!

Timeouts are still a bit weird, working on it. I implemented a super
simple retry for both http and grpc. It does some things differently but
seems to work.

Also lots of error handling & docs to clean up still.

* go mod tidy

* switch to our own interface

* add a bunch of godoc

* go mod tidy

* remove otlpclient package/dir

I like the idea but it's a lot of work to untangle it from the config,
for not much gain. Code is simpler bundled in otelcli so let's keep it
there.

* clean up diagnostics

There were some workarounds for getting errors out of the OTel stack.
Those are no longer necessary. Renamed the OtelError field to just Error
bc that's how it's used now. It's still a string because marshaling to
the error type isn't cool.

* alias proto module to a meaningful name

* add TODOs and fill out retry() some more

* fix up cases in http/UploadTraces

The success case wasn't clear. It now reads the body and closes it too.

Still need to do something with the response but that will come later.

* back out timeout hacks

This work needs to be done, but my approach was awful, and
it doesn't belong in this PR.

* pass protobuf spans as pointers

go vet started showing errors on copied protobuf spans. It wasn't really
a problem in otel-cli but go vet is right, copying a mutex is pretty
much always a bad idea.

* fix warning for unbuffered signal channel

* fill in errors

* add response and error handling to the http client

* break out http error handling for testability

add a test file, just a start

* fill in tests, fix uncovered bug

* refactor grpc error processing to a func

interim step, needs more work...

* add a SetError() method to make returns cleaner

* fill in some grpc retry code, add server wait

While filling in gRPC error handling, I added server-side recommended
delay time. This required refactoring the retryFunc signature to tunnel
that wait through to the retry function. This ended up touching http as
well because they share the retry func.

* add tests for grpc client

* go mod tidy

* fix retry docs

* remove dead code that probably caused bugs

Really need to expand the test framework to test this but it's a big
lift for another PR.