remove otlpserver.CliEvent (#221)
Refactors otel-cli's server to remove the CliEvent abstraction. It seemed like a good idea at
the time but it turns out, it was a bad decision. The protobuf spans and span events don't
look as nice but are fine to work with and ultimately using them directly is more precise.

One feature added: `otel-cli server tui` and `otel-cli server json` now support http. I was working
on that code anyways and it was easy to add along the way.

The test suite now has a separate map for checking server metadata. Previously it was
serialized to a single line, sorted lexically. It worked OK but this is cleaner, easier to read,
and easier to maintain.

Squashed commit log follows:

* fix demo exec syntax and add --verbose --fail

* refactor clievent.go away

Big commit but everything was connected so it all kinda had
to happen at once. Tests pass though :)

* remove commented code

* fix warning on tlsConf

Doesn't really belong in this PR but it was bothering me.

* remove old serverMeta passing

* re-implement server meta testing

This is better than the old way :)

* enable http server for otel-cli server tui

* fix godoc

* make server http/grpc selection more robust

Also remove a no longer needed string manipulation in the grpc code.

* make both server json and tui support the same options

* make it clear server json/tui don't support https yet