separate w3c traceparent to its own package (#225)
The traceparent code was still sitting in otlpclient after the breakup of CLI and
OTLP because it was kinda a toss-up where it should land at that stage. It is
now in its own package.

The w3c directory is there, to make it clear it's not an otel-specific thing, and as
a place to put a baggage API that's next after this PR merges.

I also made a small change to functional tests to reflect a behavior change. The
way trace ids and spans were printed in comments in non-recording mode. The
comments no longer show a zero-valued trace id in non-recording mode and
instead show the traceparent's trace & span id. In recording mode behavior is
unchanged.

No new functionality. No changes to functional tests.

Squashed commitlog follows:

* move files

* refactor w3c/traceparent so it stands alone

Still some leftover commented code that will move to protobuf span code.

* move span-dependent code to otlpclient/protobuf_span

* finish refactoring traceparent to its own package

* functionality change: just print the tp trace/span id

Making it work consistently was a pain and the code was ugly so I
decided to drop the feature where span/trace id would be zeroed if the
trace isn't recording.

* add testing & fix bugs & clean up