test headers, fix implementation of grpc headers
There was an egregious testing gap. Mea culpa.

Adds tests for headers through grpc and http.

To do that, header support had to be added to the servers, so I did
that.

That provided the data needed to augment the test harness to check
headers. Added that.

Added functional tests that exercise the headers end-to-end.

Noticed grpc.WaitForReady() while I was perusing otel collector and
grpc docs, added a TODO to look into it later. Might be useful for
otel-cli status.