move error tracking to context values, add multiple canaries to otel-cli status (#227)
The bulk of changes are function signature updates to plumb ctx all the way
through otlpclient so that errors can be tunneled out via ctx instead of globals.

Added `otel-cli status --canary-count 1 --canary-interval 10ms` functionality for
sending multiple canaries. This takes advantage of the new error tunneling to
expose errors in `otel-cli status`.

There are new methods in otlpclient to handle storing & retrieving errors in ctx.

This PR also contains the addition of otlpclient/otlp_client_null.go that implements
a non-recording client. While doing the context work I noticed I forgot to have CLI
commands stop their clients. After adding that, otel-cli started panicing on nil
pointers because the client was nil. NullClient fixes this.

Squashed commitlog follows:

* plumb ctx all the way through the OTLP clients

* add a null OTLP client

While working on context plumbing I noticed the CLI tools don't stop the
connection explicitly. I added that and started getting crashes because
the client was nil in non-recording mode. The null client fixes those
cases properly.

* more ctx work and add support for NullClient

* plumb more ctx into CLI tools, make sure they stop the client

* godoc & cleanups

Simplified some code that didn't need to be so defensive.

* save every retry's error

* add the basics of otel-cli status --keepalive

Something is broken in otlpserver so it's disconnecting after the first
span, which is a whole different bug to fix. Handily, it demonstrates
the usecase for keepalive testing.

* defend against nil errors in SaveError

* otlpclient.SendSpan() must not stop the client

This was wrong, but fine until I needed to send more than one span.

* remove duplicate deadline, re-enable grpc close

Remove duplicate deadline and especially the deferred cancel. This was
one of a couple places where otlpclient prevented itself from working
for more than one span.

Also re-enable closing the grpc connection in Stop() now that other
issues are solved.

* add --keepalive functionality to otel-cli status

Adds a first take on otel-cli status --keepalive.

* update functional test expectations for status

Previously otel-cli status didn't call ParseCliTimeout() at all so the
diagnostic wasn't set and had the zero value. Now that it's being
called, the default value of 1s gets parsed so the test needs to specify
that.

* properly reparent spans also to the trace id

* add comments, add span list to status output

Marked a couple things deprecated in status output that will be
a lot of tedious work to refactor out, and can wait.

* more comments

* move to --canary-count and --canary-interval

still more tweaking to do but this feels on the right track

* rework canary interval, factor parseDuration out for reuse

* fix Diag.ParsedTimeoutMs

* comment update

* make otel-cli status --canary-count work as expected

Default canary count bumped to 1 to keep old behavior.

* add json tags to error struct