fill in some testing gaps, fix grpc headers (#232)
While working on #227 a few gaps in testing were exposed, especially for gRPC and HTTP headers.

This PR adds functional tests for gRPC and HTTP header passing. Along the way I found the bug
in gRPC headers, introduce in #205. When I wrote the gRPC client I assumed headers would be
passed in with grpcOpts. I looked at the OTel collector, and it is passing them via grpc metadata.
So I fixed that.

Also added a test for the new ErrorList functionality. Just a single test for now but it's set up to add more.

Squashed commitlog follows:

* add a test for SaveError / GetErrorList

Not complete, but works. Set up as a table test, with just one simple
test for now.

* pass time into SaveError so it's easier to test

Also expanded test to compare error lists.

* test headers, fix implementation of grpc headers

There was an egregious testing gap. Mea culpa.

Adds tests for headers through grpc and http.

To do that, header support had to be added to the servers, so I did
that.

That provided the data needed to augment the test harness to check
headers. Added that.

Added functional tests that exercise the headers end-to-end.

Noticed grpc.WaitForReady() while I was perusing otel collector and
grpc docs, added a TODO to look into it later. Might be useful for
otel-cli status.

* actually assert against the new Headers values

also added * support to checkHeaders