refactor Config out of otlpclient (#245)
Moves the Config from otlpclient to otelcli, where it belongs. It only ended up in otlpclient
because in the first pass of splitting things up, it was easier to keep it with the client. With
this PR, otlpclient breaks dependence on Config by switching to an interface for configuration.
otelcli.Config implements that interface.

The vast majority of this change is moving code around, and adding a few methods along
way for interfaces.

No new features.

`otel-cli --otlp-blocking` is now a no-op and is labeled deprecated.

Squashed commitlog follows:

* move tlsConfig to its own file, rename to TlsConfig

* move TlsConfig to be a method attached to Config

* move SendSpan to using a config interface instead of struct

First step in splitting out config is adding getter methods to Config
and moving code to use an interface instead.

* move ParseEndpoint to Config, add GetEndpoint()

* refactor Config usage out of otlp client

Fills in OTLPConfig interface with the needed methods to break
dependence on Config so it can move out.

* remove unused function, add godoc to another

* move NewProtobufSpanWithConfig to Config, move to span_config.go

protobuf span is now clear of Config usage. This can move with Config
and can stay with it since it's all about translating otel-cli config to
a span and is useless otherwise.

* update broken tests

* move Config files to otelcli

* finish moving Config to otelcli.Config

* fix tests, fix comments

* rename files to make more sense

* fix comment

* remove dead code

Nothing was using the retry counter and Diag can't work in otlpclient
anymore so out it goes.