check for servers out of spec
In #262 it was reported that a vendor's OTLP server responded to an
http/protobuf trace upload with a JSON response. I checked with a vendor
contact and they confirmed this is not supposed to happen and means
their server is out of spec.

I might be convinced to just accept the JSON as long as it's a valid
ExportTraceServiceResponse but for now, add an error that informs users
clearly that it's the server's fault.