clean up sockets on span background end
After some tests were added, they failed on my machine. Looking into it,
it turned out that the sockfile from previous span background operations
lingered and caused subsequent tests to fail. I'm not sure why it didn't
fail in CI but it doesn't hurt to unlink the file just to be sure.

I skipped error checking on it for now, so that it doesn't have to be
careful about whether the file exists or not.