remove config.StartupTime, stop using it
Removes config.StartupTime, which used to be used to compute all the
deadlines. With this change, that variable goes away and sites use
time.Now() instead. Also, the last remaining automatic deadline is in
the retry() function. This should get removed eventually but a stub is
in place to keep things working for now.