add exec --command-timeout (#275)
Adds `otel-cli exec --timeout 1s --command-timeout 1h sleep 600` so --timeout is just for OTel and the command can run as long as it takes. This meant some reworking of timeouts throughout the code, with some remaining work to do, but this passes test and is ready to ship.

* go mod tidy

* switch to exec.CommandContext, add --command-timeout

* update help text

global exec help was outdated from before we moved away from 'sh -c'

--command-timeout still had copypasta from --timeout that was wrong

* add tests for --command-timeout

This commit will fail because the span deadline fails because it's based
on otel-cli's start time. Will fix in next commit.

* remove config.StartupTime, stop using it

Removes config.StartupTime, which used to be used to compute all the
deadlines. With this change, that variable goes away and sites use
time.Now() instead. Also, the last remaining automatic deadline is in
the retry() function. This should get removed eventually but a stub is
in place to keep things working for now.

* fix warnings in main_test.go

I switched these in some recent PR and that was the wrong decision since
test methods aren't allowed in goroutines.