fix timeouts
In my previous PR, I changed otlpclient.retry() to add a deadline when
one is missing. This would have been fine if it used the old
config.StartupTime but that's gone so I used time.Now(). The problem is,
it'll get a new time in the future for every call to retry and never
stop.

Clearly this needs more testing but for now let's fix it what wasn't
broken before.