fix exec span attrs (#316) (#318)
During previous refactors of exec for other things, the span creation got moved to the top of the function, and then later code would manipulate config.Attributes, which no longer got picked up.

Regression test added for #316.

Since this functionality was broken for a bit, the attributes are being renamed along the way to match the OTel semantic specs.

Detailed commitlog follows:

* fix exec span attrs

During previous refactors of exec for other things, the span creation
got moved to the top of the function, and then later code would
manipulate config.Attributes, which no longer got picked up.

Regression test added for #316.

Since this functionality was broken for a bit, the attributes are being
renamed along the way to match the OTel semantic specs.

Attributes are built up manually in protobuf structs because the helpers
don't support array values. This could move to a helper later but seems
fine as-is.

* remove duplicate types

* add regular expression capability for SpanData

* test new attributes, fix bugs

Adds a new RE test for generated attributes. Refactors exec.go to fix
some bugs discovered and add more attributes. Since the protobuf
expression of attributes is so verbose, it's now in its own funcs at the
bottom to keep the doExec() function readable.

* remove extra linefeed

* add CHANGELOG notes for bugfix