build smaller binaries, add version subcommand (#333)
A PR came in to add -s -w to build ldflags, which will reduce binary sizes. Since the PR only updated /Dockerfile, these will only impact locally-built binaries.

This PR adds the flags to goreleaser and github actions builds to be consistent.

Along the way I noticed that I never exposed the version information in otel-cli, so I added that too.

Also removed a content-length check in http header testing because it's not reliable and beside the point of the test it was breaking.

Detailed commitlog follows:

* set ldflags in goreleaser builds

-s -w to build smaller binaries
-X options to fill in version, commit, and date

* add version subcommand

* set -s -w and some -X ldflags on build in GHA

Ensures testing happens with a binary closer to what goreleaser will
build.

* update CHANGELOG.md for recent changes

* go mod tidy

* fix test failure due to gzip encoding

With gzip encoding enabled it's difficult to compare content-length
header because side-effects can easily break it.

The test harness now automatically ignores content-length in HTTP
headers unless the test fixture specifically sets an expected result.